refine LoopOneColumn
